requirements.txt
Django==2.2.5
django-elasticsearch-dsl==7.0.0
django-elasticsearch-dsl-drf==0.20.2
django-filter==2.2.0
django-haystack==2.8.1
django-nine==0.2.2
djangorestframework==3.10.3
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
